+++ internal/s/error.go
@@ -422,20 +422,23 @@ type RestartToleranceReached struct {
 	failedChildName        string
 	failedChildErrCount    uint32
 	failedChildErrDuration time.Duration
-	err                    error
+	sourceErr              error
+	lastErr                error
 }
 
 // NewRestartToleranceReached creates an ErrorToleranceReached record
 func NewRestartToleranceReached(
 	tolerance restartTolerance,
-	err error,
-	ch c.Child,
+	sourceCh c.Child,
+	sourceErr error,
+	lastErr error,
 ) *RestartToleranceReached {
 	return &RestartToleranceReached{
-		failedChildName:        ch.GetRuntimeName(),
+		failedChildName:        sourceCh.GetRuntimeName(),
 		failedChildErrCount:    tolerance.MaxRestartCount,
 		failedChildErrDuration: tolerance.RestartWindow,
-		err:                    err,
+		sourceErr:              sourceErr,
+		lastErr:                lastErr,
 	}
 }
 
@@ -443,8 +446,9 @@ func NewRestartToleranceReached(
 func (err *RestartToleranceReached) KVs() map[string]interface{} {
 	kvs := make(map[string]interface{})
 	kvs["node.name"] = err.failedChildName
-	if err.err != nil {
-		kvs["node.error.msg"] = err.err.Error()
+	if err.lastErr != nil {
+		kvs["node.error.source.msg"] = err.sourceErr.Error()
+		kvs["node.error.last.msg"] = err.lastErr.Error()
 		kvs["node.error.count"] = err.failedChildErrCount
 		kvs["node.error.duration"] = err.failedChildErrDuration
 	}
@@ -458,7 +462,7 @@ func (err *RestartToleranceReached) Error() string {
 // Unwrap returns the last error that caused the creation of an
 // ErrorToleranceReached error
 func (err *RestartToleranceReached) Unwrap() error {
-	return err.err
+	return err.lastErr
 }
 
 // explainLines returns a human-friendly message of the error represented as a slice
@@ -469,17 +473,25 @@ func (err *RestartToleranceReached) explainLines() []string {
 		outputLines,
 		[]string{
 			fmt.Sprintf(
-				"worker node '%s' was restarted at least %d times in a %v window; "+
-					"the last error reported was:",
+				"worker node '%s' was restarted more than %d times in a %v window.",
 				err.failedChildName,
 				err.failedChildErrCount,
 				err.failedChildErrDuration,
 			),
+			"the original error reported was:",
 		}...,
 	)
 	outputLines = append(
 		outputLines,
-		indentExplain(1, errToExplain(err.err))...,
+		indentExplain(1, errToExplain(err.sourceErr))...,
+	)
+	outputLines = append(
+		outputLines,
+		"the last error reported was:",
+	)
+	outputLines = append(
+		outputLines,
+		indentExplain(1, errToExplain(err.lastErr))...,
 	)
 	return outputLines
 }