+++ internal/s/spec.go
@@ -63,7 +63,9 @@ const (
 	// OneForOne is an Strategy that tells the Supervisor to only restart the
 	// child process that errored
 	OneForOne Strategy = iota
-	// OneForAll
+	// OneForAll is an Strategy that tells the Supervisor to restart all the
+	// siblings of a failed child process
+	OneForAll
 	// RestForOne
 )