+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+# v0.2.0
+
+* Introduce supervision restart strategy `OneForAll` (#65)
+
+* `SupervisorRestartError` explain message got enhanced with first detected error
+  on a window of time (#65)
+
 # v0.1.0 (breaking changes)
 
 * Deprecate `WithTolerance` worker option with `WithRestartTolerance` supervisor